requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.26.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
chromadb>=0.4.15
langchain-chroma>=0.1.0
sentence-transformers>=2.2.2
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5

# Document processing
PyMuPDF>=1.22.5
python-docx>=0.8.11
openpyxl>=3.1.2
langdetect>=1.0.9
fpdf>=1.7.2

# Utilities
python-dotenv>=1.0.0
